+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### HEAD
 * Add `Pry.configure` as an alternative to the current way of changing configuration options in `.pryrc` files.
+* Add `Pry::Config::Behavior#eager_load!` to add a possible workaround for issues like [#1501](https://github.com/pry/pry/issues/1501)
 * Remove Slop as a runtime dependency by vendoring v3.4 as Pry::Slop.
   People can depend on Slop v4 and Pry at the same time without running into version conflicts. ([#1497](https://github.com/pry/pry/issues/1497))
 * Fix auto-indentation of code that uses a single-line rescue ([#1450](https://github.com/pry/pry/issues/1450))